+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - improve OpenDSS -> `Network` parsing
 - expand `Network{MultiPhase}` support (for BranchFlowModel.jl)
 - add `VariableBounds`
-- add `VoltageRegulator`,  `Capacitor`, and `ShuntAdmittance` as a new bus inputs
+- add `VoltageRegulator`, `Capacitor`, and `ShuntAdmittance` as a new bus inputs
 - add `VariableContainer = Dict{Int, Dict{String, Any}}` and `VARIABLE_NAMES`
 - add `Network.var_name_map` for parsing results
 - rm all `Inputs` associated stuff